feat(src): organize code in separate files
