+++ src/builtin/echo.rs
@@ -1,0 +1,4 @@
+pub fn builtin_echo(args: Vec<String>) -> i32 {
+    println!("{}", args.join(" "));
+    0
+}
