+++ src/builtin/mod.rs
@@ -1,0 +1,2 @@
+pub mod builtin;
+pub mod echo;
